Update gh-pages
Autumn Evenings: This is a sketch I made where I first played around with the concept of inserting images into my p5.js sketches. I added an image of myself in my Halloween costume this year, as well as an image of daisies that you can move and contort. Atop everything is a gif image of golden cartoon sparkles to tie the piece together.
Flashy Spice: This was by far the most complex sketch I have made this semester, as it uses the library _p5.speak_ to use voice recognition software and listen for specific keywords in order to perform an action. I made it so that if the system hears the words "green" or "orange", it will turn those colors and repeat the words back to you.
Snow in the Night: This sketch was where I made an infinite loop of small snow particles falling from a certain point on the canvas and covered that point with contorted ellipses to appear like a snow cloud.